Update status-monitor-v2-get-started.md
In prerequisites, running the following command will fail with an error "Unable to download" if the machine does not have TLS1.2 enabled. 
Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force
This is expected as that is the reason also for having SMv1 no longer working, but it is worth to be mentioned in the doc. We got several cases on that. 
https://portal.microsofticm.com/imp/v3/incidents/details/211509874/home